Created my first custom component Also refactored Style.js into the ui folder
