+++ client/App.js
@@ -1,11 +1,11 @@
 import MaterialIcons from "@expo/vector-icons/MaterialIcons";
 import * as Font from "expo-font";
 import * as SplashScreen from 'expo-splash-screen';
-import {styles} from './Style';
 import { useCallback, useEffect, useState } from "react";
 import { ScrollView, View, Image, SectionList, Text } from "react-native";
 import { Button, CheckBox } from "react-native-elements";
-import MyButton from './ui';
+import { styles } from './ui/assets/Style';
+import { MyButton } from './ui/components/MyButton';
 
 export default function App() {
   const [appIsReady, setAppIsReady] = useState(false);
@@ -16,12 +16,12 @@ export default function App() {
       try {
         // Keep the splash screen visible while we fetch resources
         await SplashScreen.preventAutoHideAsync();
-                // Pre-load fonts, make any API calls you need to do here
+        // Pre-load fonts, make any API calls you need to do here
         await Font.loadAsync(MaterialIcons.font);
-                // Artificially delay for two seconds to simulate a slow loading
+        // Artificially delay for two seconds to simulate a slow loading
         // experience. Please remove this if you copy and paste the code!
         await new Promise(resolve => setTimeout(resolve, 2000));
-              } catch (e) {
+      } catch (e) {
         console.warn(e);
       } finally {
         // Tell the application to render
@@ -52,23 +52,23 @@ export default function App() {
   return (
     <View style={styles.container} onLayout={onLayoutRootView}>
       <Image
-                source={{
+        source={{
           width: "100%",
           height: "35%",
           uri: 'https://picsum.photos/1920/1080',
         }}
       />
-      <SectionList 
+      <SectionList
         sections={DATA}
         keyExtractor={(item, index) => item + index}
-        renderItem={({item}) => (
+        renderItem={({ item }) => (
           <View style={styles.item}>
-            <MyButton style={styles.title} title={"My Title"} />
+            <MyButton title="This is a 'MyButton'" />
           </View>
-      )}
-      renderSectionHeader={({section: {title}}) => (
-        <Text style={styles.header}>{title}</Text>
-      )}
+        )}
+        renderSectionHeader={({ section: { title } }) => (
+          <Text style={styles.header}>{title}</Text>
+        )}
       />
     </View>
   );
