+++ client/package.json
@@ -15,7 +15,8 @@
     "react-native": "0.73.2",
     "react-native-elements": "^3.4.2",
     "react-native-safe-area-context": "4.8.2",
-    "react-native-web": "~0.19.6"
+    "react-native-web": "~0.19.6",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3"
