+++ client/ui/components/MyButton.tsx
@@ -1,12 +1,8 @@
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
 import React from 'react'
 
-const MyButton = ({title}) => {
-  return (
-	<TouchableOpacity>
-	  <Text>{title}</Text>
-	</TouchableOpacity>
-  )
-}
-
-export default MyButton
+export const MyButton = ({ onPress, title }) => (
+  <TouchableOpacity onPress={onPress}>
+    <Text>{title}</Text>
+  </TouchableOpacity>
+);
